In BSLeDiscovery don’t specify main queue for posting notifications. This reverses most of commit af05fca. Revise tests to pass.
Tim Ekl of Omni Group recommended to me:
Post notification on current thread, whether it's main or a background thread.
Observers will be notified on current thread.
Let each observer decide if it will respond on current thread or not.
For example a view controller might want to get the main queue and then update UI.